Fix duplicate delete
If a rollup migration comes after a column's deletion, the
old_col.is_deleted is going to be true and new_col is going to be nil.
In this situation, the column deletion is already reported previously
when is_deleted becomes true, so don't report here.

Plus, if the column is deleted, no matter whether it's deleted in the
current version or not, don't proceed with other checks and continue
with the next column.